Reworked starters and fixed implementation mistakes
