+++ app/src/test/java/quentin/cache/GameStarterTest.java
@@ -75,6 +75,26 @@ class GameStarterTest {
                         assertTrue(
                                 output.toString()
                                         .contains(
-                                                "forward or f             : go forward one move")));
+                                                "forward or f             : go forward one move")),
+                () -> assertTrue(output.toString().contains("Available commands:\n")),
+                () ->
+                        assertTrue(
+                                output.toString()
+                                        .contains(
+                                                "  exit                     : Quits the game and"
+                                                        + " exits the program    \n")),
+                () ->
+                        assertTrue(
+                                output.toString()
+                                        .contains(
+                                                "  help                     : Shows this help      "
+                                                        + "                   \n")),
+                () ->
+                        assertTrue(
+                                output.toString()
+                                        .contains(
+                                                "  <coordinates>            : Makes a move."
+                                                        + " Examples: A1 b2 C5 (wrong examples: 5A,"
+                                                        + " 24)")));
     }
 }
